Extracted flattening logic to a different class
